adapted modules to zarr format

+++ src/electrai/dataloader/utils.py
@@ -46,7 +46,7 @@ def load_zarr(root: str | bytes | os.PathLike, index: str):
         z = zarr.open_group(str(path), mode="r")
         if "structure" not in z.attrs:
             raise KeyError(f"'structure' attribute missing from zarr store at {path}")
-        arr = np.array(z["charge_density_total"])
+        arr = np.array(z["charge_density_total"], dtype=np.float64)
         volume = json.loads(z.attrs["structure"])["lattice"]["volume"]
         return arr / volume
